+++ annonex2embl/GenerationOps.py
@@ -20,7 +20,7 @@ from Bio.SeqFeature import ExactPosition, FeatureLocation, CompoundLocation
 ###############
 
 __author__ = 'Michael Gruenstaeudl <m.gruenstaeudl@fu-berlin.de>'
-__copyright__ = 'Copyright (C) 2016-2019 Michael Gruenstaeudl'
+__copyright__ = 'Copyright (C) 2016-2020 Michael Gruenstaeudl'
 __info__ = 'annonex2embl'
 __version__ = '2019.10.16.1700'