+++ scripts/annonex2embl_launcher_CLI.py
@@ -2,9 +2,12 @@
 # -*- coding: utf-8 -*-
 
 if __name__ == '__main__':
-    
+
     import sys, os
     sys.path.append(os.path.join(os.path.dirname(os.path.dirname(__file__)), 'annonex2embl'))
-    from annonex2embl import CLIOps
+    try:
+        from annonex2embl import CLIOps
+    except:
+        import CLIOps
 
     CLIOps.start_annonex2embl()